requirements.txt
flet==0.24.*
pandas
httpx
openpyxl
requests
python-dotenv
aiohttp